Refactor CreateProjectStage to use ReleaseDate instead of StartDate and update related bindings
