+++ src/Angor/Avalonia/AngorApp/Sections/Founder/CreateProject/Stages/CreateProjectStage.cs
@@ -3,12 +3,12 @@ using ReactiveUI.Validation.Extensions;
 using ReactiveUI.Validation.Helpers;
 using Zafiro.UI.Commands;
 
-namespace AngorApp.Sections.Founder.CreateProject;
+namespace AngorApp.Sections.Founder.CreateProject.Stages;
 
 public partial class CreateProjectStage : ReactiveValidationObject, ICreateProjectStage
 {
     [Reactive]
-    private DateTimeOffset? startDate;
+    private DateTime? releaseDate;
 
     [Reactive]
     private double? percent;
@@ -18,9 +18,7 @@ public partial class CreateProjectStage : ReactiveValidationObject, ICreateProje
         Remove = ReactiveCommand.Create(() => remove(this)).Enhance();
         this.ValidationRule(stage => stage.Percent, x => x is >= 1, "Percent must be greater than 1");
         this.ValidationRule(stage => stage.Percent, x => x != null, "Enter a percentage");
-        this.ValidationRule(stage => stage.StartDate, x => x != null, "Enter a date");
-        StartDate = DateTimeOffset.Now;
-        Percent = 1;
+        this.ValidationRule(stage => stage.ReleaseDate, x => x != null, "Enter a relase date");
     }
 
     public IEnhancedCommand Remove { get; }
