+++ src/Angor/Avalonia/AngorApp/Sections/Founder/CreateProject/Stages/ICreateProjectStage.cs
@@ -1,11 +1,11 @@
 using ReactiveUI.Validation.Abstractions;
 using Zafiro.UI.Commands;
 
-namespace AngorApp.Sections.Founder.CreateProject;
+namespace AngorApp.Sections.Founder.CreateProject.Stages;
 
 public interface ICreateProjectStage : IReactiveObject, IValidatableViewModel
 {
     public double? Percent { get; set; }
-    public DateTimeOffset? StartDate { get; set; }
+    public DateTime? ReleaseDate { get; set; }
     public IEnhancedCommand Remove { get; }
 }
