+++ src/Angor/Avalonia/AngorApp/Sections/Founder/CreateProject/Stages/StagesView.axaml.cs
@@ -1,8 +1,4 @@
-using Avalonia;
-using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
-
-namespace AngorApp.Sections.Founder.CreateProject;
+namespace AngorApp.Sections.Founder.CreateProject.Stages;
 
 public partial class StagesView : UserControl
 {
